% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parglo.maxlik.R, R/pglo.R
\name{Generalized-Logistic}
\alias{Generalized-Logistic}
\alias{parglo.maxlik}
\alias{pglo}
\alias{pglo}
\title{Generalized Logistic distribution function.}
\usage{
pglo(x, para)

parglo.maxlik(x, ini)

pglo(x, para)
}
\arguments{
\item{x}{vector of quantiles for which to evaluate the PDF.}

\item{ini}{a vector of initial values of the parameters to be fit.}

\item{para}{a list of parameters of a generalized Logistic 
distribution function, as yield by function \code{parglo}.}
}
\value{
A vector of cumulative probabilities corresponding to the 
quantiles in \code{x}.
}
\description{
Cumulative distribution function of the 
Generalized Logistic probability distribution function, 
and maximum likelihood fitting function.

See parglo.maxlik
}
\details{
This functions are used internally by \code{spei} 
and \code{spi} and are supposed to never be needed by the 
regular user. Initial values for maximum likelihood estimation
can be provided by \code{parglo}.

See parglo.maxlik
}
\references{
S.M. Vicente-Serrano, S. Beguería, J.I. López-Moreno. 2010. 
A Multi-scalar drought index sensitive to global warming: 
The Standardized Precipitation Evapotranspiration Index – SPEI. 
\emph{Journal of Climate} \bold{23}: 1696, DOI: 10.1175/2009JCLI2909.1.
}
\author{
Santiago Begueria
}
