% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgi.R
\name{sgi}
\alias{sgi}
\title{Groundwater drought index}
\usage{
sgi(data, scale, kernel = list(type = "rectangular", shift = 0),
  distribution = "GNorm", fit = "ub-pwm", na.rm = FALSE,
  ref.start = NULL, ref.end = NULL, x = FALSE, params = NULL, ...)
}
\description{
`sgi`: generalized normal distribution (pelgno) used here.
}
