% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spei2.R
\name{spei2}
\alias{spei2}
\title{Modified SPEI}
\usage{
spei2(
  data,
  scale,
  kernel = list(type = "rectangular", shift = 0),
  distribution = "log-Logistic",
  fit = "ub-pwm",
  na.rm = FALSE,
  ref.start = NULL,
  ref.end = NULL,
  x = FALSE,
  params = NULL,
  ...,
  date = NULL
)
}
\arguments{
\item{data}{a vector, matrix or data frame with time ordered values 
of precipitation (for the SPI) or of the climatic balance 
precipitation minus potential evapotranspiration (for the SPEI).}

\item{scale}{an integer, representing the time scale at which 
the SPEI / SPI will be computed.}

\item{kernel}{optional, a list defining the type of kernel used 
for computing the SPEI / SPI at scales higher than one. Defaults 
to unshifted rectangular kernel.}

\item{distribution}{optional, name of the distribution function 
to be used for computing the SPEI / SPI (one of 'log-Logistic', 
'Gamma' and 'PearsonIII'). Defaults to 'log-Logistic' for \code{spei}, 
and to 'Gamma' for \code{spi}.}

\item{fit}{optional, name of the method used for computing the 
distribution function parameters (one of 'ub-pwm', 'pp-pwm' and 
'max-lik'). Defaults to 'ub-pwm'.}

\item{na.rm}{optional, a logical value indicating whether NA values 
should be stripped from the computations. Defaults to FALSE, i.e. 
no NA are allowed in the data.}

\item{ref.start}{optional, starting point of the reference period 
used for computing the index. Defaults to NULL, indicating that the 
first value in data will be used as starting point.}

\item{ref.end}{optional, ending point of the reference period used 
for computing the index. Defaults to NULL, indicating that the last 
value in data will be used as ending point.}

\item{x}{optional, a logical value indicating wether the data used 
for fitting the model should be kept. Defaults to FALSE.}

\item{params}{optional, an array of parameters for computing the 
spei. This option overrides computation of fitting parameters.}

\item{...}{other possible parameters.}

\item{date}{(optional) A Date Vector with the same length as date.}
}
\description{
Modified SPEI
}
\author{
Dongdong Kong (Sun Yat-sen Univ)
}
\keyword{internal}
